Fix customer discovery pipeline losing ~30% of customers from sales output
Switch to smaller contiguous SalesOrderNumberKey

+++ src/facts/sales/sales_worker/schemas.py
@@ -120,7 +120,7 @@ def build_worker_schemas(
     ]
 
     order_fields = [
-        pa.field("SalesOrderNumber", pa.int64()),
+        pa.field("SalesOrderNumber", pa.int32()),
         pa.field("SalesOrderLineNumber", pa.int64()),
     ]
     delta_fields = [pa.field("Year", pa.int16()), pa.field("Month", pa.int16())]
@@ -159,7 +159,7 @@ def build_worker_schemas(
     #   - Detail remains line-level for product/pricing/shipping facts
     # ---------------------------------------------------------------------
     detail_fields = [
-        pa.field("SalesOrderNumber", pa.int64()),
+        pa.field("SalesOrderNumber", pa.int32()),
         pa.field("SalesOrderLineNumber", pa.int64()),
         pa.field("ProductKey", pa.int64()),
         pa.field("DueDate", pa.date32()),
@@ -174,7 +174,7 @@ def build_worker_schemas(
     detail_schema = pa.schema(detail_fields + delta_fields) if is_delta else pa.schema(detail_fields)
 
     header_fields = [
-        pa.field("SalesOrderNumber", pa.int64()),
+        pa.field("SalesOrderNumber", pa.int32()),
         pa.field("CustomerKey", pa.int64()),
         pa.field("StoreKey", pa.int64()),
         pa.field("SalesPersonEmployeeKey", pa.int64()),
